📝(review) add documentation for code reviews
When new developers are joining the team, it is important for them to
get to know the practices in use. Code reviews are an exercise for which
the team wants to keep good practices. Adapted from an article from
TailorDev, the goal and the good practices of code reviews have been
documented.